requirements.txt
# NOTE: This file is maintained for convenience only.
# The canonical dependency list is in pyproject.toml.
# Preferred install: pip install -e ".[all]"

# Core dependencies
openai
python-dotenv
fire
httpx
rich
tenacity
prompt_toolkit
pyyaml
requests
jinja2
pydantic>=2.0
PyJWT[crypto]
debugpy

# Web tools
firecrawl-py
parallel-web>=0.4.2

# Image generation
fal-client

# Text-to-speech (Edge TTS is free, no API key needed)
edge-tts

# Optional: For cron expression parsing (cronjob scheduling)
croniter

# Dashboard TUI
textual>=0.50.0

# Optional: For messaging platform integrations (gateway)
python-telegram-bot[webhooks]>=22.6
discord.py>=2.0
aiohttp>=3.9.0
